Change from 'cp' to 'rsync' when doing 'make www'
The motivation here is to delete things that are no longer used
when they go stale, something that we were not doing before.

This involved a refactor to the Makefile so that both the preview and
www steps now use the same `ln` commands to link in the static web
content that's outside of hugo.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>