Replace ChapelCon CFP news item with Registration one; add forgotten image
This adds a news item for ChapelCon registration and manually removes
the one for the CFP by making its date 1000 years ago.  I didn't just
remove it because I'd like it to show up in a list of all news items
once https://github.com/chapel-lang/chapel-www/issues/34 and
https://github.com/chapel-lang/chapel-www/issues/33 are implemented
(the first so that it doesn't show up on the front page; the second so
that it can be seen in the comprehensive view).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>