Rename 'ChapelCon' directory to 'chapelcon'
Because of the new `chapelcon/index.html` file I've added as a replacement
for ChapelCon.html, we now have a conflict between the old 'ChapelCon'
directory and the new 'chapelcon' one which doesn't cause problems on
case-insensitive platforms like my Mac, but does cause problems on the
real website (or rather, when developing the real website on a Mac and
then pushing it to the live site).

This PR renames the 'ChapelCon' directory to 'chapelcon' and updates
existing links to reflect the new naming while also setting up a
redirect for the old capitalization.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>